backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Yash Prakash <yashprakash13@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^0.21.1"
uvicorn = {extras = ["standard"], version = "^0.20.0"}
fastapi = "^0.89.1"
tweepy = {extras = ["async"], version = "^4.12.1"}


[tool.poetry.group.dev.dependencies]
black = "^22.12.0"
pre-commit = "^3.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
